backend/requirements.txt
fastapi
uvicorn[standard]
openai
google-generativeai
anthropic
python-dotenv
PyPDF2
pdf2image
requests
python-multipart
lxml
beautifulsoup4 # Added for HTML parsing
PyMuPDF # Added for PDF processing (fitz module)

# XML Parsing
xmltodict

# FHIR/EHR Processing
fhir.resources>=7.1.0

# RTF Parsing
striprtf

# Database and ORM
prisma==0.15.0          # Prisma ORM for database access
asyncpg==0.30.0         # PostgreSQL async driver
psycopg2-binary==2.9.10 # PostgreSQL driver

# Security and Authentication
bcrypt==4.3.0           # Password hashing
python-jose[cryptography]==3.4.0  # JWT tokens
email-validator==2.2.0  # Email validation for Pydantic
